use XDG_CACHE_HOME instead of ~/popcon.cache
people generally don't appreciate non-hidden files in their home
directories, and the .cache directory (which is the typical location of
XDG_CACHE_HOME) is already widely used (26 entries in my case) and
specified in [1].

although it's relatively trivial to implement the relevant parts
directly here, i think it's cleaner to take the funktionality from
python-xdg, which is a new dependency.

the precise location to store the popcon data is
$XDG_CACHE_HOME/popcon/debian. this was chosen instead of
$XDG_CACHE_HOME/popcon as other distros (ubuntu) might want to adapt
this tool and upgrading from a situation where something is a file to
one where that thing is a directory sounds quite painful.

[1] http://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html