filebased timestamping
importing the cache file can be quite slow, and timestamp comparison
originally happens only after the file is loaded.

the system is here changed to mtime based comparison, removing the
timestamp from the pickled format. to catch this and future format
changes, the timestamp is not only compared to the expiry condition but
also to the source file's timestamp.

as a result, the code flow slightly changed (less flow through exception
handling). errors reading the file are still ignored, but now trigger a
warning.