module level data caching
in keeping with the way the module is configured (using global
variables), a package global cache is introduced that allows subsequent
popcon lookups to use the unpickled data instead of opening the file and
unpickling every time (which is time consuming).

the cache is invalidated based on time using the same rules as the file
cache in case the module is used in long-running processes.