atomic file creation
in the previous version, the dump files were created in place. as a
result, a process reading the cache while another writes it might get
corrupted data (although i guess it would rather raise an exception; not
good either way, although that case would in the end only cause a
warning).

the new code creates a temporary file and then renames it to the proper
name. i'm not an expert in posix semantics, but that should pretty much
solve the problem of race conditions.

a typical scenario in which this could have been a real problem is a web
server running several processes. when starting up, all of them have the
same cache expiry time. if after that time requests hit different
processes in short intervals, the first opens the cache file for
writing and truncates it; the second one sees the recent cache file and
tries to load from it (and fails). with the new code, worst thing to
happen is that both download their new copies independently and rename
them to the cache file. (they could be synchronized to wait for each
other, but i don't want to implement ipc stuff right now, especially
when thinking of cache directories shared among different servers.)

note that it's quite realistic to be caught in that race condition as
pickling the dictionary takes several seconds on my box.

concerning clear code: this is a bit complicated to retain compatibility
with python2.5 (which will be in stable until squeeze is released). when
python2.5 compatibility is dropped, it can be put more clearly using
delete=False.